refactor(tx-builder): migrate tx/builder/field-types.js to ts
